Point feature: Fixing typos and moving Point3D to separate class.
